Correções

+++ src/main/java/com/example/demo/model/ItemCarrinho.java
@@ -24,7 +24,7 @@ import javax.persistence.OneToOne;
 public class ItemCarrinho {
     private Long id;
     private Produto produto;
-    private double quantidade;
+    private Long quantidade;
     
     private Carrinho carrinho;
 
@@ -48,11 +48,11 @@ public class ItemCarrinho {
         this.produto = produto;
     }
     
-    public double getQuantidade() {
+    public Long getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(double quantidade) {
+    public void setQuantidade(Long quantidade) {
         this.quantidade = quantidade;
     }
